% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{fiRstore_auth}
\alias{fiRstore_auth}
\title{Authorize fiRstore}
\usage{
fiRstore_auth(
  email = gargle::gargle_oauth_email(),
  path = NULL,
  scopes = "https://www.googleapis.com/auth/datastore",
  cache = gargle::gargle_oauth_cache(),
  use_oob = gargle::gargle_oob_default(),
  token = NULL
)
}
\arguments{
\item{email}{\itemize{
\item Google identity of the user
}}

\item{path}{\itemize{
\item JSON key file for a service account user.
}}

\item{scopes}{\itemize{
\item Additional permissions associated with the token
}}

\item{cache}{\itemize{
\item OAuth token cache.
}}

\item{use_oob}{\itemize{
\item Are you using the out-of-band authentication
}}

\item{token}{\itemize{
\item a prepared token (not sure why this is here,
maybe I already have a token ... see \code{\link[gargle:token_fetch]{gargle::token_fetch()}} for more info)
}}
}
\value{
Returns an invisible token.
}
\description{
Authorizes fiRstore to access documents in Google Firestore.  This function
is a wrapper around \code{\link[gargle:token_fetch]{gargle::token_fetch()}}.

The package does not supply a default application, so you will need to provide
an application by calling fiRstore_auth_config(app = your_app).  To create
an application call \code{\link[httr:oauth_app]{httr::oauth_app()}}
}
