% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestoreClient.R
\name{fiRstore_get}
\alias{fiRstore_get}
\title{Get a document from Google Cloud Firestore}
\usage{
fiRstore_get(project, database = "(default)", collection, document)
}
\arguments{
\item{project}{- the GCP project}

\item{database}{- The firestore database default = (default)}

\item{collection}{- Document collection in the data base..}

\item{document}{- Document that you are retrieving.}
}
\value{
the json document.
}
\description{
Wraps the [GCP Firestore get REST API](https://firebase.google.com/docs/firestore/reference/rest/v1/projects.databases.documents/get)
}
