% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestoreClient.R
\name{fiRstore_get_document}
\alias{fiRstore_get_document}
\title{Get a document from Google Cloud Firestore}
\usage{
fiRstore_get_document(project, database = "(default)", collection, document)
}
\arguments{
\item{project}{- the GCP project}

\item{database}{- The firestore database default = (default)}

\item{collection}{- Document collection in the data base..}

\item{document}{- Document that you are retrieving.}
}
\value{
the json document.
}
\description{
Get a document from Google Cloud Firestore
}
