% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firestoreClient.R
\name{fiRstore_runQuery}
\alias{fiRstore_runQuery}
\title{Search for documents in GCP cloud Firestore}
\usage{
fiRstore_runQuery(project, database = "(default)", collection, query)
}
\arguments{
\item{project}{the GCP project}

\item{database}{The firestore database default = (default)}

\item{collection}{Document collection in the data base..}

\item{query}{JSON text}
}
\value{
JSON text containing the results of the query
}
\description{
Runs a query and returns the documents
Wraps the \href{https://firebase.google.com/docs/firestore/reference/rest/v1/projects.databases.documents/runQuery}{GCP Firestore runQuery REST API}

An example query would look like this

\if{html}{\out{<div class="sourceCode">}}\preformatted{\{"structuredQuery":
     \{"from": [\{
         "collectionId":"people",
         "allDescendants":false\}],
      "where":\{
        "fieldFilter":\{
           "field":\{"fieldPath":"FirstName"\},
           "op":"EQUAL",
           "value":\{"stringValue":"Daniel"\}
           \}
         \}
     \}
 \}
}\if{html}{\out{</div>}}
}
