% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{fiRstore_auth_config}
\alias{fiRstore_auth_config}
\title{Configure application for fiRstore}
\usage{
fiRstore_auth_config(app)
}
\arguments{
\item{app}{- an httr oauth application

This is change in the future as gargle transitions from
httr to httr2.}
}
\description{
Configure application for fiRstore
}
